### 2026-03-27 #763 @mpadge
@smwindecker I couldn't work out why your submission template had everything styled like headers, then I realised it was because the YAML front-matter requires an empty line between last variable (`Language: ...`) and the closing YAML line. Without that space, it interprets `line\n---` as a markdown header. That also unfortunately entirely mucks up the ability of the bot to parse the variables it needs to get things going. That is, you've currently got:
```
Language: en
---
```
But you need to have:
```
Language: en

---
```
Then an additional problem is that your HTML variables have been removed. I suggest that you copy the header part of a fresh template across and update yours here, making sure to keep all HTML variables like this:
```
Submitting Author Github Handle: @github_handle
```
The bot works entirely on the HTML variable defs (``), so those have to be kept in place. And of course that also means the full fix above ends up like this:
```
Language: en

---
```

So can you please edit the template above to fix all of these issues, and then try `@ropensci-review-bot check package`? Sorry for any confusion here, and thanks for submitting!

---

And I've just updated our system to deliver hopefully much more informative error messages in cases like those. For yours, you'd see an initial message that the YAML spacer has been removed, and after fixing that, you'd see additional error messages advising to reinstate HTML variables. You can try with the bot if you want to have fun getting told off by a bot 😆 ... or you can just fix like I've advised here and trigger the checks again 👍
