### 2026-03-30 #763 @smwindecker
Submitting Author Name: Saras Windecker
Version submitted: https://github.com/acefa-hubs/EpiStrainDynamics/releases/tag/v3.0_rOpenSci-submission
Badge grade: gold (select one)

---

```
Package: EpiStrainDynamics
Title: Infer temporal trends of multiple pathogens
    c(person("Saras", "Windecker", email = "saras.windecker@gmail.com",
      person("Oliver", "Eales", role = "aut",
      person("James", "McCaw", role = "aut",
      person("Freya", "Shearer", role = "aut",
      person("Milad", "Kharratzadeh", role = "ctb")
    )
Description: 'EpiStrainDynamics' is a statistical framework developed for inferring temporal trends of multiple pathogens from routinely collected surveillance data.
License: Apache License 2.0
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
Biarch: true
    bayesplot,
    cli,
    Rcpp (>= 0.12.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.18.1),
    rstantools (>= 2.4.0),
    timetk,
    tsibble,
    viridis
    BH (>= 1.66.0),
    Rcpp (>= 0.12.0),
    RcppEigen (>= 0.3.3.3.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.18.1),
    StanHeaders (>= 2.18.0)
SystemRequirements: GNU make
    piggyback,
    tibbletime,
    units,
    xts,
    zoo
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in [issue#749](#749)

EpiStrainDynamics is a statistical modelling framework capable of inferring case trends of multiple pathogens. Estimating the temporal trends in infectious disease activity is crucial for monitoring disease spread and the impact of interventions. The target audience for this package is researchers and decision-makers who use surveillance data to monitor disease trends.

The modelling framework presented here has scientific applications related to different surveillance systems covering multiple pathogens (eg. influenza, SARS-CoV-2, dengue), scenarios (seasonal epidemics, non-seasonal epidemics, pandemic emergence), and temporal reporting resolutions (weekly, daily). This methodology is applicable to a wide range of pathogens and surveillance systems.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

This software expands upon an existing modelling algorithm, developed carefully for the test case of multiple pathogens modelled in an epidemiological context. There is no other existing relevant software for the earlier versions of these models but there are published manuscripts with sections of stan model code made public.

[Eales O, de Oliveira ML, Page AJ, et al. Dynamics of competing SARS-CoV-2 variants during the Omicron epidemic in England. Nat Commun. 2022;13(1):1-11. 10.1038/s41467-022-32096-4](https://www.nature.com/articles/s41467-022-32096-4)
[Eales O, Ainslie KEC, Walters CE, et al. Appropriately smoothing prevalence data to inform estimates of growth rate and reproduction number. Epidemics. 2022;40:100604. 10.1016/j.epidem.2022.100604](https://pubmed.ncbi.nlm.nih.gov/35780515/)

NA

-    What grade of badge are you aiming for? ([gold](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges))

1.	Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories: We demonstrate, in detail, performance of our models against multiple statistical standards. In particular we use simulated data to demonstrate our models' algorithmic scaling, scale appropriateness, scale recovery, and robustness to noise and random seeds. We address these standards (G5.7, G5.9, G5.9a, G5.9b, BS7.3, BS7.4, BS7.4a) via the vignette titled "Algorithmic Scaling".
2.	Have a demonstrated generality of usage beyond one single envisioned use case: This software is designed to be applicable to different surveillance systems covering multiple pathogens (eg. influenza, SARS-CoV-2, dengue), scenarios (seasonal epidemics, non-seasonal epidemics, pandemic emergence), and temporal reporting resolutions (weekly, daily).
3.	Internal aspects of package structure and design: Given the broad extent of potential use cases, the data intake steps in particular were designed with extra care to allow a user to specify a number of different modelling scenarios with ease. The model set up used in the data intake then flows clearly into the modelling and post-processing metrics calculations steps so that that user only needs to specify scenarios at the outset. These design choices we believe make the package of superior design.

> I struggled to get autotest to run.

## Use of Generative AI

- [x] Generative AI tools were used to produce some of the material in this submission.

If so, please describe usage, and include links to any relevant aspects of your repository. See [our blog post](https://ropensci.org/blog/2026/02/26/ropensci-ai-policy/) for background. (Explicit advice is not yet included in our _Dev Guide_; we are hoping to update very soon, and ask your cooperation and transparency in the meantime.)

Claude was used to generate first drafts of some of the test suite. In particular the [list column tests](https://github.com/acefa-hubs/EpiStrainDynamics/blob/main/tests/testthat/test-list_columns.R), the [units pkg support tests](https://github.com/acefa-hubs/EpiStrainDynamics/blob/main/tests/testthat/test-units-pkg-support.R), and the [verbosity tests](https://github.com/acefa-hubs/EpiStrainDynamics/blob/main/tests/testthat/test-verbosity.R).
