### 2026-03-30 #763 @smwindecker
@mpadge I tried!! Looks like we are in a circular pattern that it won't let me because it doesn't recognize me as author1. Please let me know if I should start a fresh issue.

In case you're curious, I think this happneed because I copied from my first submission, which lost some of the meta info. Thanks for your patience as I apparently trigger every kind of error.
