### 2026-03-30 #763 @mpadge
Thanks @smwindecker but you still hadn't restored the submission template metadata properly. I've edited now and fixed it all up. The essential empty line before YAML close was missing, and most variables still didn't have the HTML defs, most importantly including `@smwindecker`. With that, you can't be automatically identified as author here, so the bot won't let you do anythning. Please try again now and it should work.
