### 2026-04-25 #763 @adamhsparks
Hi @smwindecker, a few comments as I start looking for reviewers. Mostly stylistic/linting for you to consider, no showstoppers.

1. The {goodpractice} check has rightly pointed out that you import packages as a whole, which is usually undesirable. Is there any way to avoid this?
2. I note that you've used `\dontrun{}` in your examples, from experience this can come back to bite you on CRAN, you may want to consider using `\examplesIf{ interactive()}` instead.
3. You've used a mix of """ and "'" quotes, using """ is more standard in R, but using just one except where you need to use `paste()` or the like would be cleaner.
4. Place the arguments that do not have a default value before the ones that do.
5. Consider running the tests in parallel for faster package checks, put `Config/testthat/parallel: true` in your DESCRIPTION file.
6. I note that there is "vignettes/Using-EpiStrainDynamics.Rmd.orig" and "vignettes/Using-EpiStrainDynamics.Rmd" but I don't see how they are related, is there a compile script that is called that is missing?
