### 2026-05-03 #763 @maurolepore
I see @adamhsparks has posted his initial comments and is searching for reviewers. Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
