### 2026-07-17 624bc4c
revise review version of NEWS to highlight original features, link the issues that changes addressed.
closes #41
