+++ NEWS.md
@@ -1,14 +1,34 @@
 # EpiStrainDynamics (development version)
-* Modified the plot code to explicitly call all ggplot2 functions and not import the whole package. 
-* Removed uses of \dontrun{} in the examples. Slow examples use @examplesIf interactive() and other cases use alternatives, eg. @examplesIf rlang::is_installed("xts") where xts is used only in the example but not imported by the package.
-* Standardised use of quotation marks style. 
-* Placed the arguments that do not have a default value before the ones that do.
-* Added Config/testthat/parallel: true to DESCRIPTION file.
-* Added additional tests for mcmc validations. 
 
-EpiStrainDynamics 0.0.1.0000 (2026-03-27)
-=========================
+## Minor improvements
 
-### New features
+* Plot functions now call `ggplot2` functions explicitly rather than importing
+  the whole package (#35).
+* Standardised quotation-mark style (#37).
+* Reordered function arguments so those without defaults precede those with
+  defaults (#38).
+* Enabled parallel testing with `Config/testthat/parallel: true` (#39).
+* Added tests for MCMC validation.
 
-  * submitted to ROpenSci
+## Documentation
+
+* Replaced `\dontrun{}` in examples with `@examplesIf interactive()` for slow
+  examples and guards such as `@examplesIf rlang::is_installed("xts")` where a
+  package is used only in an example (#36).
+
+# EpiStrainDynamics 0.0.1 (2026-03-27)
+
+Initial version submitted to rOpenSci for peer review
+(ropensci/software-review#763).
+
+## New features
+
+* Random-walk and p-spline models for inferring temporal trends of single and
+  multiple (including subtyped) pathogens, fitted via Stan.
+* Model construction and fitting via `construct_model()` and `fit_model()`,
+  with `smoothing_structure()` and `dispersion_structure()` helpers.
+* Post-processing generics `incidence()`, `growth_rate()`, `Rt()`, and
+  `proportion()`, each with a `plot()` method.
+* `diagnose_model()` for MCMC convergence diagnostics.
+* Bundled `influenza` and `sarscov2` example datasets.
+* "Using EpiStrainDynamics" and "Algorithmic Scaling" vignettes.
