### 2026-07-17 6b8609a
revise documentation for sarscov2 dataset - correct the number of columns listed and the descriptions of column names
#56
