+++ R/data.R
@@ -4,30 +4,22 @@
 #' 2020 - 2022 retrieved from the UK Health Security Agency's data dashboard.
 #' downloaded data describing the daily number of variants detected by
 #' collection date. The data classified all sequences based on ‘major lineage
-#' calls’. We considered 11 groupings of the variants based on their major
+#' calls’. We considered groupings of the variants based on their major
 #' lineage calls or other, consisting of all lineages with a designation not
 #' consistent with any of the major lineage calls.
 #'
 #' @format ## `sarscov2`
-#' A data frame with 830 rows and 15 columns:
+#' A data frame with 830 rows and 6 columns:
 #' \describe{
 #'   \item{date}{Date, between 23 September 2020 and 31 December 2022}
 #'   \item{cases}{Numeric, daily number of cases}
-#'   \item{B.1.177}{Integer, daily number of cases of B.1.177}
-#'   \item{B.1.1.7}{Integer, daily number of cases of B.1.1.7 (Alpha variant)}
-#'   \item{B.1.617.2}{Integer, daily number of cases of B.1.617.2 (Delta variant)}
-#'   \item{BA.1}{Integer, daily number of cases of BA.1 (Omicron BA.1 variant)}
-#'   \item{BA.2}{Integer, daily number of cases of BA.2 (Omicron BA.2 variant)}
-#'   \item{BA.2.75}{Integer, daily number of cases of BA.4 (Omicron BA.4 variant)}
-#'   \item{BA.4}{Integer, daily number of cases of BA.5 (Omicron BA.5 variant)}
-#'   \item{BA.5}{Integer, daily number of cases of BA.2.75 (Omicron BA.2.75 variant)}
-#'   \item{BQ.1}{Integer, daily number of cases of BQ.1 (Omicron BQ.1 variant)}
-#'   \item{XBB}{Integer, daily number of cases of XBB (a recombinant of omicron
-#'      sub-variants)}
-#'   \item{Other}{Integer, daily number of cases of all lineages with a
-#'      designation not consistent with any of the major lineage calls}
-#'   \item{total}{}
-#'   \item{t}{}
+#'   \item{alpha}{Integer, daily number of cases of B.1.1.7 (Alpha variant)}
+#'   \item{delta}{Integer, daily number of cases of B.1.617.2 (Delta variant)}
+#'   \item{omicron}{Integer, daily number of cases of BA.1, BA.2, BA.2.75,
+#'      BA.4, BA.5, BQ.1 (Omicron variants)}
+#'   \item{other}{Integer, daily number of cases of B.1.177, XBB (a recombinant
+#'      of omicron sub-variants), and all other lineages with a designation
+#'      not consistent with any of the major lineage calls}
 #' }
 #' @source <https://ukhsa-dashboard.data.gov.uk/covid-19-archive-data-download;
 #'   https://datadryad.org/dataset/doi:10.5061/dryad.hx3ffbgm2>
