### 2026-07-17 004ebb9
revise also the documentation for influenza dataset. was not caught in review but also failed to correctly reflect structure of datatset
closes #56
