+++ R/data.R
@@ -42,7 +42,7 @@
 #' causes which we do not have data to determine).
 #'
 #' @format ## `influenza`
-#' A data frame with 830 rows and 15 columns:
+#' A data frame with 426 rows and 7 columns:
 #' \describe{
 #'   \item{ili}{Integer, daily total number of cases of influenza-like illness}
 #'   \item{week}{Date, between 1 January 2012 to week starting 1 March 2020}
