### 2026-07-17 ebdab41
Revise contribution statement to link the CONTRIBUTING.md and add detail on scope of welcome contributions (closes #58).
Add note indicating the package is being actively maintained (closes #51).
