### 2026-07-17 feb33dd
Document C++ toolchain requirement for building from source in README
There is no CRAN release yet, so installing from GitHub always compiles
the bundled Stan models from source. Add setup guidance per OS, closes #60.
