### 2026-07-17 1eb9251
Clarify contribution scope in README
