### 2026-07-17 fdbc94c
Add citations to intro paragraph and a bridging paragraph before Step 1
Cite the two Eales et al. papers this package's modelling framework
builds on, and add a short transition explaining that building a model
means calling construct_model() then fit_model(), before diving into
the step-by-step example.

Closes #57, closes #59
