### 2026-07-17 2515319
Document tau and phi, and their priors_provided fields, in smoothing/dispersion structures
Explain what tau and phi represent and how they affect smoothness/
overdispersion, grounded in the smoothing parameter's definition in
Eales et al. 2022 (Epidemics). Document the priors_provided field in
@return for both functions, which was previously unmentioned.

Closes #53
