### 2026-07-17 46c2fc2
Expand diagnose_model() documentation and fix example bug
Document the expected class of fitted_model, the elements of the
returned list, and link to definitions of R-hat and effective sample
size for readers unfamiliar with these MCMC diagnostics. Also fix the
roxygen example, which called the nonexistent diagnose(fit) instead of
diagnose_model(fit).

Closes #54
