### 2026-07-17 05ca1bb
Use roxygen [] cross-links for function references in docs
Backtick-wrapped function names in construct_model()'s and plot()'s
documentation don't render as clickable links. Convert the references
to actual function calls (random_walk(), p_spline(), single(),
multiple(), subtyped(), smoothing_structure(), dispersion_structure(),
incidence(), growth_rate(), Rt(), proportion()) to use [] instead.

Closes #55
