### 2026-07-17 3e0edc6
Bump minimum R version to 4.1.0 and test it in CI
The base pipe (|>), used in compute_multi_pathogen(), requires R
4.1.0 and cannot be backported — the prior R (>= 3.5) floor in
DESCRIPTION was already inaccurate. Add the declared minimum to the
R-CMD-check.yaml matrix so it's actually tested rather than asserted.

Closes #43
