+++ DESCRIPTION
@@ -22,8 +22,8 @@ Language: en
 RoxygenNote: 7.3.3
 Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
 Biarch: true
-Depends: 
-    R (>= 3.5)
+Depends:
+    R (>= 4.1.0)
 Imports: 
     bayesplot,
     cli,
