### 2026-07-17 7896fe4
Convert simple stop() calls to cli::cli_abort() (part of #49)
Converts the 6 plain message-based stop() calls in diagnose_model.R,
validation-smoothing.R, and validation-utils.R to cli::cli_abort(),
matching the {.arg}/{.val}/{.cls} style already used elsewhere in the
validation code. Update the handful of tests that asserted the old
exact wording, since cli's markup changes quoting/formatting.

Refs #49
