### 2026-07-17 68ee554
Convert fit_model()'s custom error re-throws to cli::cli_abort() (closes #49)
Extract a shared abort_fit_error() helper that raises the
EpiStrainDynamics.fit.error condition via cli::cli_abort() instead of
stop(), collapsing 12 near-identical blocks across the 6 fit_model.*
methods into one-line calls. Verified end-to-end (not just unit-level)
that the catchable contract is preserved: class, $message,
$constructed_model, and conditionCall() all survive, matching what
test-verbosity.R asserts.

Also loosen one test's error-message regex: it asserted the literal
Stan exception text, but on this 
[…]
