+++ tests/testthat/test-verbosity.R
@@ -149,10 +149,13 @@ test_that("fit_model catches Stan failures and provides informative output", {
   # Set standata to NULL - this will cause Stan to fail
   mod$standata <- NULL
 
-  # Now it should throw an actual error
+  # Now it should throw an actual error. The exact message depends on
+  # whether this rstan/Stan toolchain raises a catchable R error during
+  # sampler creation or just prints and returns an incomplete stanfit
+  # object, so match on the comment/label both paths share.
   expect_error(
     fit_model(mod, n_iter = 500, n_chain = 1, verbose = FALSE),
-    "Stan sampling failed"
+    "Stan sampling failed|failed to create the sampler"
   )
 })
 
