### 2026-07-17 b16d73b
update news with first batch of peer review reponse changes
