### 2026-07-17 1be49e4
Stop redefining plot() as a generic; use methods on the base generic (closes #48)
plot() was previously redefined and exported by the package, which
would shadow graphics::plot for any other loaded package's plot()
usage while EpiStrainDynamics is attached. Remove the standalone
generic and rely on base R's existing plot() generic instead --
plot.incidence/growth_rate/Rt/proportion are registered as S3 methods
via NAMESPACE's S3method() entries regardless of whether the package
defines its own generic. Move validate_class_inherits() into each of
the four methods since there's no shared generic
[…]
