+++ NAMESPACE
@@ -31,7 +31,6 @@ export(growth_rate)
 export(incidence)
 export(multiple)
 export(p_spline)
-export(plot)
 export(proportion)
 export(random_walk)
 export(single)
