+++ articles/algorithmic-scaling.md
@@ -295,10 +295,10 @@ knitr::kable(scaling_results, digits = 3,
 
 | method      | pathogen_structure | scaling_exponent | r_squared |
 |:------------|:-------------------|-----------------:|----------:|
-| p_spline    | multiple           |            1.009 |     0.997 |
-| p_spline    | single             |            0.840 |     0.990 |
-| random_walk | multiple           |            0.958 |     0.999 |
-| random_walk | single             |            0.883 |     0.984 |
+| p_spline    | multiple           |            1.001 |     0.997 |
+| p_spline    | single             |            0.851 |     0.992 |
+| random_walk | multiple           |            0.962 |     0.999 |
+| random_walk | single             |            0.879 |     0.981 |
 
 Scaling exponents: time complexity approximately O(n^exponent) {.table}
 
@@ -803,7 +803,7 @@ on data similar to its intended use case.
 ``` r
 
 sessionInfo()
-#> R version 4.6.0 (2026-04-24)
+#> R version 4.6.1 (2026-06-24)
 #> Platform: x86_64-pc-linux-gnu
 #> Running under: Ubuntu 24.04.4 LTS
 #> 
@@ -824,39 +824,39 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] dplyr_1.2.1               ggplot2_4.0.3            
-#> [3] EpiStrainDynamics_0.0.1.1
+#> [1] dplyr_1.2.1                  ggplot2_4.0.3               
+#> [3] EpiStrainDynamics_0.0.1.9000
 #> 
 #> loaded via a namespace (and not attached):
 #>  [1] tidyselect_1.2.1      viridisLite_0.4.3     timeDate_4052.112    
-#>  [4] farver_2.1.2          viridis_0.6.5         loo_2.9.0            
+#>  [4] farver_2.1.2          viridis_0.6.5         loo_2.10.0           
 #>  [7] S7_0.2.2              fastmap_1.2.0         digest_0.6.39        
 #> [10] rpart_4.1.27          timechange_0.4.0      lifecycle_1.0.5      
 #> [13] StanHeaders_2.32.10   survival_3.8-6        magrittr_2.0.5       
-#> [16] compiler_4.6.0        rlang_1.2.0           sass_0.4.10          
-#> [19] tools_4.6.0           yaml_2.3.12           data.table_1.18.4    
+#> [16] compiler_4.6.1        rlang_1.3.0           sass_0.4.10          
+#> [19] tools_4.6.1           yaml_2.3.12           data.table_1.18.4    
 #> [22] knitr_1.51            labeling_0.4.3        htmlwidgets_1.6.4    
-#> [25] pkgbuild_1.4.8        RColorBrewer_1.1-3    withr_3.0.2          
+#> [25] pkgbuild_1.4.8        RColorBrewer_1.1-3    withr_3.0.3          
 #> [28] purrr_1.2.2           desc_1.4.3            nnet_7.3-20          
-#> [31] grid_4.6.0            stats4_4.6.0          timetk_2.9.1         
+#> [31] grid_4.6.1            stats4_4.6.1          timetk_2.9.1         
 #> [34] xts_0.14.2            future_1.70.0         inline_0.3.21        
 #> [37] globals_0.19.1        scales_1.4.0          MASS_7.3-65          
 #> [40] anytime_0.3.13        cli_3.6.6             rmarkdown_2.31       
[…]
