### 2026-07-20 f9903cd
Add a print() method for EpiStrainDynamics.model objects (closes #47)
Printing a constructed model previously dumped the full data and
standata list to the console. Add print.EpiStrainDynamics.model(),
registered as an S3 method on the base print() generic (not a
redefinition of it, per the same pattern used for plot() in #48),
showing method, pathogen structure, pathogen names, observation
count, and DOW flag, plus a head() of the underlying tsibble.
