+++ man/print.EpiStrainDynamics.model.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/construct_model.R
+\name{print.EpiStrainDynamics.model}
+\alias{print.EpiStrainDynamics.model}
+\title{Print method for constructed EpiStrainDynamics models}
+\usage{
+\method{print}{EpiStrainDynamics.model}(x, ...)
+}
+\arguments{
+\item{x}{an \code{EpiStrainDynamics.model} object, as returned by
+\code{\link[=construct_model]{construct_model()}}}
+
+\item{...}{further arguments passed to or from other methods (unused)}
+}
+\value{
+\code{x}, invisibly
+}
+\description{
+Prints a concise summary instead of dumping the full data and standata
+list to the console.
+}
+\examples{
+mod <- construct_model(
+  method = random_walk(),
+  pathogen_structure = single(
+    data = sarscov2,
+    case_timeseries = "cases",
+    time = "date"
+  )
+)
+print(mod)
+}
