+++ NAMESPACE
@@ -22,6 +22,7 @@ S3method(plot,Rt)
 S3method(plot,growth_rate)
 S3method(plot,incidence)
 S3method(plot,proportion)
+S3method(print,EpiStrainDynamics.model)
 export(Rt)
 export(construct_model)
 export(diagnose_model)
