### 2026-07-20 2647e9b
Set pkgdown math-rendering to mathjax to fix broken Rt() equation on the site (refs #44)
Sangeeta's "the equation does not render" comment was about the pkgdown
site, not `?Rt` locally -- these are two separate rendering pipelines
(R's own HTML help viewer vs. pkgdown's static site), so it's expected
that one could be broken while the other looks fine.

Traced it to a real bug: pkgdown 2.1.0 changed the default math
renderer from MathJax to MathML ("low-dependency... lowest fidelity"
per pkgdown's own docs), and Rt()'s equation is complex enough (nested
\frac, a \sum with sub/superscript limit
[…]
