### 2026-07-21 c0ed933
Document gi_dist's validation requirements and scope in Rt() (closes #44)
Add to the "Generation interval considerations" bullets: gi_dist is
validated for structural properties only (vectorized, non-negative,
finite), does not need to sum to 1 since Rt() normalizes internally
via g_a, and validation does not assess epidemiological plausibility
of the shape -- supplying an accurate distribution is the user's
responsibility.
