### 2026-07-21 8ac81e5
Add the precompile script for the vignette and include contribution instructions (closes #46)
We already incorporated the .Rmd.orig and knitted .Rmd files. We now remove the *.R to track the precompile script itself. We also add instructions to the contributing.md to clarify how to make revisions to the vignette.
