### 2026-07-22 9deddb1
Fix the indexing so that the mean is a scalar per observation (closes #42)
Revise the .stan file, recompile, and add instructions for revisions to stan models to the contributing.md.
