### 2026-07-23 5926be7
Resolve faulty figure rendering by revising the fi path in the vignette itself and the knit options in the compile file (closes #45)
knitr::knit() resolves fig.path relative to the working directory at call time, not relative to the input/output file. Since fig.path in the vignette is set relative to the vignette itself (e.g. "figures/..."), base.dir must be set to vignettes/ here so generated figures land in vignettes/figures/ instead of <wherever precompile.R was run from>/figures/.
