+++ vignettes/precompile.R
@@ -1,4 +1,5 @@
 # Pre-compiled vignettes that are slow to run
-# Must manually move image files to vignettes/ after knit
+
+knitr::opts_knit$set(base.dir = "vignettes")
 knitr::knit("vignettes/Using-EpiStrainDynamics.Rmd.orig",
             output = "vignettes/Using-EpiStrainDynamics.Rmd")
