+++ vignettes/Using-EpiStrainDynamics.Rmd
@@ -1,6 +1,8 @@
 ---
 title: "Using EpiStrainDynamics"
-output: rmarkdown::html_vignette
+output:
+  rmarkdown::html_vignette:
+    pandoc_args: ["--from", "markdown-implicit_figures"]
 vignette: >
   %\VignetteIndexEntry{Using EpiStrainDynamics}
   %\VignetteEngine{knitr::rmarkdown}
@@ -32,7 +34,7 @@ method = p_spline(spline_degree = 3,  # example value for spline degree
 ## Pathogen structure
 
 There are three main types of pathogen structure available to model: `single()`, `multiple()`, and `subtyped()`.
-These functions require the name of the dataset itself and the column names for different data elements. These can also ingest timeseries class objects, which would then make the `time` column specification optional. The models are not currently equipped to handle missing values or irregular date sequences. Timeseries checks are run internally on data frame or any time of timeseries class input that is handled by the [`tsbox` package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html).
+These functions require the name of the dataset itself and the column names for different data elements. These can also ingest timeseries class objects, which would then make the `time` column specification optional. Any timeseries class input that is handled by the [`tsbox` package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html) is acceptable. The models are not currently equipped to handle missing values or irregular date sequences.
 
 The `single()` pathogen structure is the simplest and models a single pathogen timeseries. The name of the dataframe is passed to argument `data`, the name of the column with total case data is passed to `case_timeseries`, and the name of the column of time data is passed to `time` (optional if input dataset is a timeseries class). It can be specified as follows, illustrated using data provided with the package `sarscov2`:
 
@@ -90,7 +92,9 @@ smoothing_params = smoothing_structure(
 
 ## Dispersion parameters
 
-The argument `dispersion_params` allows users to set a prior for the overdispersion parameter of the negative binomial likelihood for the case timeseries. It is specified using `dispersion_structure()` as below:
+The argument `dispersion_params` allows users to set a prior for the overdispersion parameter, `phi`, of the negative binomial likelihood for the case timeseries. This parameter controls how much the reported case count can vary from day to day around the model's expected value. Even if the model's smooth trend exactly captured the true expected number of cases on a given day, the actual reported count would still vary, since it depends on which patients happened to seek care, get tested, or have their test processed and reported that day. A negative binomial (rather than Poisson) likelihood is used because real surveillance counts are almost always more variable than a simple random-sampling process would predict, a pattern known as overdispersion. Smaller `phi` allows more day-to-day variation in reported counts; larger `phi` keeps them closer to the trend.
+
+It is specified using `dispersion_structure()` as below:
 
 ```
 dispersion_params = dispersion_structure(
@@ -100,11 +104,17 @@ dispersion_params = dispersion_structure(
 
 ## Pathogen noise
 
[…]
