### 2026-07-23 f09e181
Rename plot methods' first argument from df to x to match the generic
R CMD check flags an S3 generic/method consistency warning when a
method's first argument name doesn't match the generic's -- after #48
removed our own redefined plot() generic in favor of base R's
function(x, y, ...), our four methods still named their first
argument df. Rename to x throughout plot.incidence/growth_rate/Rt/
proportion and their bodies; no call sites used df= by name, so this
is a safe, non-breaking rename. Verified signatures now match and
validation/rendering still work.

Also update the "plot() validates 
[…]
