### 2026-07-23 812337e
add the rest of the ROpenSci review comments
