+++ articles/Using-EpiStrainDynamics.md
@@ -58,11 +58,10 @@ and
 These functions require the name of the dataset itself and the column
 names for different data elements. These can also ingest timeseries
 class objects, which would then make the `time` column specification
-optional. The models are not currently equipped to handle missing values
-or irregular date sequences. Timeseries checks are run internally on
-data frame or any time of timeseries class input that is handled by the
-[`tsbox`
-package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html).
+optional. Any timeseries class input that is handled by the [`tsbox`
+package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html)
+is acceptable. The models are not currently equipped to handle missing
+values or irregular date sequences.
 
 The
 [`single()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/single.md)
@@ -146,8 +145,20 @@ standard deviation of the prior on tau can also be specified, as below:
 ### Dispersion parameters
 
 The argument `dispersion_params` allows users to set a prior for the
-overdispersion parameter of the negative binomial likelihood for the
-case timeseries. It is specified using
+overdispersion parameter, `phi`, of the negative binomial likelihood for
+the case timeseries. This parameter controls how much the reported case
+count can vary from day to day around the model’s expected value. Even
+if the model’s smooth trend exactly captured the true expected number of
+cases on a given day, the actual reported count would still vary, since
+it depends on which patients happened to seek care, get tested, or have
+their test processed and reported that day. A negative binomial (rather
+than Poisson) likelihood is used because real surveillance counts are
+almost always more variable than a simple random-sampling process would
+predict, a pattern known as overdispersion. Smaller `phi` allows more
+day-to-day variation in reported counts; larger `phi` keeps them closer
+to the trend.
+
+It is specified using
 [`dispersion_structure()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/dispersion_structure.md)
 as below:
 
@@ -157,15 +168,36 @@ as below:
 
 ### Pathogen noise
 
-Whether to include noise between individual pathogens as well as the
-observation noise is specified as a logical (`TRUE` or `FALSE`) to the
-argument `pathogen_noise`.
+By default, the model assumes that on any given day, the relative mix of
+pathogens follows exactly the smooth underlying trends, and that any
+noisiness in the observed composition (e.g. how many positive tests were
+pathogen A vs. B) comes only from the day-to-day reporting variation
+described with `phi`, above.
+
+It is possible to add an extra source of variability, allowing the
+*true* day-to-day mix of pathogens to jitter around the smooth trend,
+not just the observed counts. This matters because real pathogens don’t
+necessarily move in lockstep with a single shared epidemic curve; one
+pathogen can genuinely spike or dip relative to the others for reasons
+the smooth trend alone doesn’t capture.
+
+The argument `pathogen_noise` refers to whether to include this noise
+between individual pathogens and can be specified as a logical (`TRUE`
[…]
