### 2026-07-23 8adae05
Apply styler to R/, tests/, and vignettes/ (refs #50)
Run styler::style_pkg()-equivalent formatting across R/, tests/testthat/,
and the vignette sources (Using-EpiStrainDynamics.Rmd.orig plus the two
standalone article vignettes), addressing the bulk of lintr's mechanical
findings: line length, indentation, spacing around operators/commas/
braces/parentheses.
