+++ tests/testthat/test-validation-utils.R
@@ -21,57 +21,93 @@ test_that("validate_positive_whole_number() returns invisible NULL", {
 })
 
 test_that("validate_positive_whole_number() rejects non-numeric inputs", {
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number("5", "test_param"),
-               "Argument test_param must be numeric")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(TRUE, "test_param"),
-               "Argument test_param must be numeric")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(list(5), "test_param"),
-               "Argument test_param must be numeric")
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number("5", "test_param"),
+    "Argument test_param must be numeric"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(TRUE, "test_param"),
+    "Argument test_param must be numeric"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(list(5), "test_param"),
+    "Argument test_param must be numeric"
+  )
 })
 
 test_that("validate_positive_whole_number() rejects multiple values", {
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(c(1, 2, 3), "test_param"),
-               "Argument test_param must be a single value")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(c(5, 10), "test_param"),
-               "Argument test_param must be a single value")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(integer(0), "test_param"),
-               "Argument test_param must be a single value")
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(c(1, 2, 3), "test_param"),
+    "Argument test_param must be a single value"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(c(5, 10), "test_param"),
+    "Argument test_param must be a single value"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(integer(0), "test_param"),
+    "Argument test_param must be a single value"
+  )
 })
 
 test_that("validate_positive_whole_number() rejects non-finite values", {
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(NA, "test_param"),
-               "Argument test_param must be numeric")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(Inf, "test_param"),
-               "Argument test_param must be a finite number")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(-Inf, "test_param"),
-               "Argument test_param must be a finite number")
-  expect_error(EpiStrainDynamics:::validate_positive_whole_number(NaN, "test_param"),
-               "Argument test_param must be a finite number")
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(NA, "test_param"),
+    "Argument test_param must be numeric"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(Inf, "test_param"),
+    "Argument test_param must be a finite number"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(-Inf, "test_param"),
+    "Argument test_param must be a finite number"
+  )
+  expect_error(
+    EpiStrainDynamics:::validate_positive_whole_number(NaN, "test_param"),
+    "Argument test_param must be a finite number"
+  )
 })
 
 test_that("validate_positive_whole_number() rejects non-whole numbers", {
[…]

[…]
