### 2026-07-23 896ad74
Drop explicit return() on final function statements (refs #50)
Removes explicit return() from the 29 lintr return_linter hits in R/,
all confirmed as the last statement in their enclosing function (or,
in one case, a nested tryCatch error-handler lambda) rather than early
returns, matching tidyverse style's preference for implicit return.
Verified with a full smoke test across construct_model, fit_model,
diagnose_model, growth_rate, Rt, incidence, smoothing_structure,
dispersion_structure, and the method constructors.
