### 2026-07-23 37fff46
Convert sapply() to vapply() for type-safe chain seed extraction (refs #50)
Replace the 6 identical sapply(fit_object@stan_args, function(x) x$seed)
calls across fit_model.rw/ps/rw_single/ps_single/rw_multiple/ps_multiple
with vapply(..., integer(1)), matching the confirmed integer type of
x$seed. The equivalent sapply() in stanmodels.R is left untouched --
it's rstantools-generated ("do not edit by hand") and returns a list
of stanmodel S4 objects, which isn't a vapply case anyway.

Verified with a real multi-chain fit that chain_seeds still returns
the correct length and type.
