+++ R/fit_model.R
@@ -211,7 +211,7 @@ fit_model.rw_subtyped <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
@@ -284,7 +284,7 @@ fit_model.ps_subtyped <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
@@ -358,7 +358,7 @@ fit_model.rw_multiple <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
@@ -431,7 +431,7 @@ fit_model.ps_multiple <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
@@ -504,7 +504,7 @@ fit_model.rw_single <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
@@ -577,7 +577,7 @@ fit_model.ps_single <- function(constructed_model,
     constructed_model = constructed_model,
     mcmc_info = list(
       seed = seed,
-      chain_seeds = sapply(fit_object@stan_args, function(x) x$seed),
+      chain_seeds = vapply(fit_object@stan_args, function(x) x$seed, integer(1)),
       init = rstan::get_inits(fit_object),
       n_chain = n_chain,
       n_iter = n_iter,
