### 2026-07-23 70f13be
Fix genuine object_usage_linter findings, leave glue-interpolation false positives (closes #50)
Delete path_names in proportion.R -- assigned but never referenced
anywhere in the file (confirmed via full-file search).

Drop the redundant `out <-` assignment in compute_multi_pathogen()
(post_fit_calc.R), returning the list directly instead. It worked
before only because R's `<-` invisibly returns the assigned value;
explicit return of the value is more robust against someone later
appending a line after it.

The other 6 object_usage_linter hits across construct_model.R and
validation-utils.R ar
[…]
