### 2026-07-23 a5e3be3
Expand growth_rate/incidence/proportion documentation to match Rt() (closes #66)
Sangeeta noted Rt() was documented in more detail than comparable metric functions. growth_rate() had no symbol breakdown for its formula and no
cross-reference to Rt() (positive growth rate corresponds to Rt > 1 and vice versa); incidence()'s day-of-week adjustment formula
introduced three symbols explained only in a prose sentence rather than itemized like Rt()'s own extra formula; proportion() didn't spell out what the numerator/denominator subscripts refer to.
