### 2026-07-23 70472cd
update NEWS.md with the finalized set of changes from this review cycle
Changes include adding issues #51, #52, and #62, which had been omitted; #43 was miscategorized under Documentation instead of Minor
improvements and didn't mention the CI matrix change; #44's MathJax fix wasn't mentioned separately from its documentation half.
