+++ NEWS.md
@@ -4,7 +4,10 @@
 
 * Fixed the negative-binomial likelihood in the random-walk single-pathogen
   model when `dow_effect = TRUE` (#42).
-* Corrected the `proportion()` / `plot()` example in the vignette and assured figures aren't missing (#45).
+* Corrected the `proportion()` / `plot()` example in the vignette and fixed
+  the underlying figure path so plots render correctly (#45).
+* Fixed the `Rt()` equation failing to render on the pkgdown site by setting
+  MathJax as the math renderer (#44).
 
 ## Minor improvements
 
@@ -18,21 +21,40 @@
 * Added a `print()` method for `EpiStrainDynamics.model` objects (#47).
 * Standardised error handling on `cli::cli_abort()` (#49).
 * `plot()` is no longer redefined as a generic (#48).
+* Increased the minimum R version to 4.1.0, matching actual base pipe usage,
+  and added it to the R-CMD-check.yaml test matrix (#43).
+* Fixed all lintr-identified issues (#50): styling (line length, indentation,
+  spacing) via `styler`, dropped explicit `return()` on final statements,
+  converted `sapply()` to `vapply()` for type-safe extraction, and removed
+  genuine dead code flagged by unused-variable checks.
 
 ## Documentation
 
 * Replaced `\dontrun{}` in examples with `@examplesIf interactive()` for slow
   examples and guards such as `@examplesIf rlang::is_installed("xts")` where a
   package is used only in an example (#36).
-* Increased R version dependency in line with base pipe usage (#43).
-* Expanded documentation for `Rt()` (#44), and`smoothing_structure()` and
-  `diagnose_model()`, add cross-linked references where applicable (#53, #54, #55).
-* Committed `precompile.R` for vignette and included revision instructions (#46).
-* Clarified contribution and maintenance guidelines, installation from source
-  instructions, and add citations to readme (#51, #57, #58, #59, #60).
+* Expanded documentation for `Rt()`, including its `gi_dist` argument (#44).
+* Expanded documentation for `smoothing_structure()` and `diagnose_model()`
+  (#53, #54).
+* Expanded documentation for `growth_rate()`, `incidence()`, and
+  `proportion()` to match the level of detail in `Rt()`, including symbol
+  breakdowns for their adjustment formulas and, for `growth_rate()`, its
+  relationship to `Rt()` (#66).
+* Added roxygen `[]` cross-links for referenced functions throughout the
+  documentation (#55).
+* Committed `precompile.R` for the vignette and included revision
+  instructions in CONTRIBUTING.md (#46).
+* Clarified contribution and maintenance guidelines, added installation-
+  from-source instructions, and added citations to the README (#51, #57,
+  #58, #59, #60).
 * Corrected documentation of both `sarscov2` and `influenza` datasets (#56).
-* Added explanations of "pathogen noise" and the day-of-week effect (#61), and a
-  COVID-19 example (#63), to the vignette.
+* Added explanations of "pathogen noise" and the day-of-week effect (#61), a
+  link to the relevant timeseries class (#62), and a COVID-19 example (#63),
+  to the vignette.
+* Clarified the G1.6 srr standard statement to note that, although
+  performance/scaling is evaluated in the vignettes, this does not
+  constitute an explicit comparison against alternative implementations
+  (#52).
 
 # EpiStrainDynamics 0.0.1 (2026-03-27)
 
