### 2026-07-23 1a98cf4
Explicitly import %||% from rlang instead of relying on base R's version
R-CMD-check on R 4.1.0 (the CI matrix entry added for #43) failed with
"could not find function '%||%'" in construct_model() and
post_fit_calc(), both in R CMD check (examples, tests) and at the
package level. Base R only gained its own %||% in R 4.4.0; this code
has been implicitly relying on that since sometime before this review,
undetected because every other R version in the CI matrix is >= 4.4.
rlang (already an Imports dependency) has its own %||% and has for
years, so import it explicitly via @importFrom rather th
[…]
