+++ NAMESPACE
@@ -60,6 +60,7 @@ importFrom(ggplot2,theme_bw)
 importFrom(ggplot2,xlab)
 importFrom(ggplot2,ylab)
 importFrom(purrr,map2)
+importFrom(rlang,"%||%")
 importFrom(rlang,.data)
 importFrom(rstan,extract)
 importFrom(rstan,sampling)
