+++ articles/algorithmic-scaling.md
@@ -28,71 +28,75 @@ simulate_sir <- function(n_days, R0, recovery_rate = 0.1, I0 = 0.01) {
   S <- numeric(n_days)
   I <- numeric(n_days)
   R <- numeric(n_days)
-  
+
   S[1] <- 1 - I0
   I[1] <- I0
   R[1] <- 0
-  
+
   beta <- R0 * recovery_rate
-  
+
   for (t in 2:n_days) {
-    dS <- -beta * S[t-1] * I[t-1]
-    dI <- beta * S[t-1] * I[t-1] - recovery_rate * I[t-1]
-    dR <- recovery_rate * I[t-1]
-    
-    S[t] <- max(0, S[t-1] + dS)
-    I[t] <- max(0, I[t-1] + dI)
-    R[t] <- min(1, R[t-1] + dR)
+    dS <- -beta * S[t - 1] * I[t - 1]
+    dI <- beta * S[t - 1] * I[t - 1] - recovery_rate * I[t - 1]
+    dR <- recovery_rate * I[t - 1]
+
+    S[t] <- max(0, S[t - 1] + dS)
+    I[t] <- max(0, I[t - 1] + dI)
+    R[t] <- min(1, R[t - 1] + dR)
   }
-  
+
   return(I)
 }
 
 # Function to generate epidemic-like data with pathogen succession
 generate_test_data <- function(n_timepoints, n_pathogens = 4, baseline_cases = 100) {
-  dates <- seq.Date(from = as.Date("2020-01-01"), 
-                    by = "day", 
-                    length.out = n_timepoints)
-  
+  dates <- seq.Date(
+    from = as.Date("2020-01-01"),
+    by = "day",
+    length.out = n_timepoints
+  )
+
   # Simulate pathogen dynamics with succession
   # Each pathogen has different R0 and timing
   pathogen_dynamics <- matrix(0, nrow = n_timepoints, ncol = n_pathogens)
-  
+
   # Pathogen 1: Early peak, moderate R0
   pathogen_dynamics[, 1] <- simulate_sir(n_timepoints, R0 = 2.5, I0 = 0.02)
-  
+
   # Pathogen 2: Mid-season peak, higher R0
   offset2 <- round(n_timepoints * 0.3)
   if (offset2 < n_timepoints) {
-    pathogen_dynamics[offset2:n_timepoints, 2] <- 
+    pathogen_dynamics[offset2:n_timepoints, 2] <-
       simulate_sir(n_timepoints - offset2 + 1, R0 = 3.0, I0 = 0.01)
   }
-  
+
   # Pathogen 3: Late peak, lower R0
   offset3 <- round(n_timepoints * 0.6)
   if (offset3 < n_timepoints) {
-    pathogen_dynamics[offset3:n_timepoints, 3] <- 
+    pathogen_dynamics[offset3:n_timepoints, 3] <-
       simulate_sir(n_timepoints - offset3 + 1, R0 = 2.0, I0 = 0.015)
   }
-  
+
   # Pathogen 4: Background/endemic, low level
-  pathogen_dynamics[, 4] <- 0.05 * sin(seq(0, 2*pi, length.out = n_timepoints)) + 0.1
-  
+  pathogen_dynamics[, 4] <- 0.05 * sin(seq(0, 2 * pi, length.out = n_timepoints)) + 0.1
+
   # Normalize to proportions
   row_sums <- rowSums(pathogen_dynamics)
   proportions <- pathogen_dynamics / row_sums
-  
+
   # Generate total cases with epidemic-like pattern
   epidemic_trend <- rowSums(pathogen_dynamics) * baseline_cases
   total_cases <- rpois(n_timepoints, lambda = epidemic_trend)
-  
+
[…]
