### 2026-07-23 c26c392
pkgcheck was failing since it was being triggered before the CI checks has passed. change to manual trigger.
