+++ .github/workflows/pkgcheck.yaml
@@ -4,9 +4,6 @@ concurrency:
   cancel-in-progress: true
 on:
   workflow_dispatch:
-  push:
-    branches:
-      - main
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
