### 2026-07-23 e9191ed
Update NEWS.md
