### 2026-07-24 bbb7da8
fix Windows R CMD check failing intermittently during Stan compilation
The Windows job was failing at a different point in the build on each
run (once after all six Stan models had compiled, once after only two),
which is the signature of virtual memory exhaustion rather than a
deterministic code bug: GitHub's Windows runners ship a small default
pagefile, and compiling six Eigen/StanHeaders/Boost-heavy Stan models
back to back can exceed it even when physical RAM looks sufficient.
Enlarging the pagefile on the Windows job resolves this without
touching compiler flags or reducing optimisation.
