+++ .github/workflows/R-CMD-check.yaml
@@ -33,6 +33,13 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - uses: al-cheb/configure-pagefile-action@v1.4
+        if: runner.os == 'Windows'
+        with:
+          minimum-size: 16GB
+          maximum-size: 16GB
+          disk-root: "C:"
+
       - uses: r-lib/actions/setup-pandoc@v2
 
       - uses: r-lib/actions/setup-r@v2
