### 2026-07-27 d760fca
serialise Stan model compilation and lower optimisation on Windows
The pagefile enlargement alone didn't resolve the intermittent Windows
R CMD check failures. Re-examining the latest failing log showed
"cc1plus.exe: out of memory" during compilation, with g++ invocations
for different stanExports_*.cc files only milliseconds apart -- a
sign that Rtools was compiling multiple heavy
Eigen/Boost/StanHeaders-templated translation units concurrently,
which the pagefile can't help with since it's a per-process compiler
memory failure, not swap starvation.

Force MAKEFLAGS=-j1 on the Windows job so 
[…]
