+++ .github/workflows/R-CMD-check.yaml
@@ -40,6 +40,11 @@ jobs:
           maximum-size: 16GB
           disk-root: "C:"
 
+      - name: Force serial compilation on Windows
+        if: runner.os == 'Windows'
+        shell: bash
+        run: echo "MAKEFLAGS=-j1" >> "$GITHUB_ENV"
+
       - uses: r-lib/actions/setup-pandoc@v2
 
       - uses: r-lib/actions/setup-r@v2
