### 2026-07-27 362c5c5
fix Makevars.win overriding the wrong variable (CXXFLAGS vs CXX17FLAGS)
The previous Makevars.win set CXXFLAGS to lower optimisation to -O1,
but had no effect: since src/Makevars sets CXX_STD = CXX17, R's build
system compiles with CXX17/CXX17FLAGS, not CXX/CXXFLAGS, so the
override was silently ignored. The latest Windows CI failure log
confirms this -- the actual g++ invocation still showed -O2. Override
CXX17FLAGS instead.
