+++ NEWS.md
@@ -12,14 +12,15 @@
   versions before 4.4.0 with "could not find function '%||%'", by
   explicitly importing `%||%` from `rlang` instead of relying on base R's
   version, which only exists from R 4.4.0 (related to #43).
-* Fixed intermittent R CMD check failures on Windows CI ("cc1plus.exe: out
-  of memory" / "Build process failed"), caused by Rtools compiling the six
-  bundled Stan models' heavy Eigen/Boost/StanHeaders template
-  instantiations in parallel and exceeding the runner's available memory,
-  by forcing serial compilation (`MAKEFLAGS=-j1`) and lowering
-  optimisation to `-O1` for Windows specifically via a new
-  `src/Makevars.win`. Also enlarged the Windows runner's pagefile as an
-  additional safeguard.
+* Fixed intermittent R CMD check failures on Windows CI ("Build process
+  failed" during compilation of the six bundled Stan models' heavy
+  Eigen/Boost/StanHeaders template instantiations), by lowering
+  optimisation from R's Windows default (`-O2`) to `-O1` via a new
+  `src/Makevars.win`, substantially reducing peak compiler memory. Since
+  `CXX_STD` is `CXX17`, this overrides `CXX17FLAGS` specifically (not
+  `CXXFLAGS`, which R does not use for this package's compilation). Also
+  forced serial compilation (`MAKEFLAGS=-j1`) and enlarged the Windows
+  runner's pagefile as additional safeguards.
 
 ## Minor improvements
 
