### 2026-07-27 f031cf2
override -O2 via PKG_CXXFLAGS, since Makeconf clobbers CXX17FLAGS
Neither of the last two attempts (setting CXXFLAGS, then CXX17FLAGS)
had any effect -- the failing Windows log's compile line was
byte-identical to before in both cases, still showing -O2. Per
"Writing R Extensions", R's own etc/Makeconf is included *after*
Makevars.win, so its plain assignment to CXX17FLAGS silently
overwrites whatever we set there.

PKG_CXXFLAGS is different: R's build rule only reads it, Makeconf
never reassigns it, which is why it's the documented mechanism for
this. Append -O1 there instead, positioned last
[…]
