### 2026-07-27 6675cf1
use override directive so -O1 survives Makeconf's later assignment
Appending -O1 to PKG_CXXFLAGS also had no effect: the latest failing
log confirms it lands in the command line *before* CXX17FLAGS's -O2
(R's ALL_CXXFLAGS template always positions PKG_CXXFLAGS ahead of
CXX17FLAGS), so -O2 still wins as the last -O flag regardless of file
inclusion order.

GNU Make's `override` directive is immune to being reset by any later
plain assignment -- from an included makefile, the environment, or the
command line -- which is exactly what's needed to survive Makeconf's
subsequent CXX17FLAGS assignment
[…]
