### 2026-07-27 cbbff17
suppress compiler warnings on Windows to test pipe-stall hypothesis
-O1 is confirmed applied via the override directive (the failing log
now shows only -O1 in the compile line, no -O2 anywhere), but the
build still fails at the same point with no literal out-of-memory
message anywhere in the log, and the whole 6-model compile completes
in ~12 seconds -- implausibly fast for genuinely heavy Stan template
compilation, and unchanged by MAKEFLAGS=-j1.

This points away from pure compiler memory exhaustion and toward the
R-side build monitor losing sync with the compiler's stdout: each
stanExports_
[…]
