### 2026-07-27 5fd7d06
reinstall rstan from source on Windows -- this was the real fix
Every attempt so far (pagefile, MAKEFLAGS=-j1, -O1, -w) was chasing
the wrong signal: "Error in proc$get_built_file(): Build process
failed" looked like a compile failure, but with -w finally cutting
the log down from 230,000+ lines to ~21,000, the real error a few
lines above it is now impossible to miss:

  Error in inDL(...) : unable to load shared object
  'D:/a/_temp/Library/rstan/libs/x64/rstan.dll':
  LoadLibrary failure: The specified module could not be found.

All six Stan models were compiling and linking into
EpiStrain
[…]
