+++ NEWS.md
@@ -12,19 +12,18 @@
   versions before 4.4.0 with "could not find function '%||%'", by
   explicitly importing `%||%` from `rlang` instead of relying on base R's
   version, which only exists from R 4.4.0 (related to #43).
-* Fixed intermittent R CMD check failures on Windows CI ("Build process
-  failed" during compilation of the six bundled Stan models' heavy
-  Eigen/Boost/StanHeaders template instantiations), via a new
-  `src/Makevars.win`: lowered optimisation from R's Windows default
-  (`-O2`) to `-O1`, and suppressed compiler warnings (`-w`), since the
-  Eigen/Boost/StanHeaders headers generate an enormous volume of
-  template-instantiation warnings that may overwhelm the small stdout
-  pipe buffers on Windows. This uses an `override CXX17FLAGS = ...`
-  directive, since R's own `etc/Makeconf` is included *after*
-  `Makevars.win` and a plain `=` assignment there would otherwise
-  silently overwrite a plain reassignment made here. Also forced serial
-  compilation (`MAKEFLAGS=-j1`) and enlarged the Windows runner's
-  pagefile as additional safeguards.
+* Fixed intermittent R CMD check failures on Windows CI. The reported
+  error ("Build process failed") was misleading: all six bundled Stan
+  models were actually compiling and linking successfully, but the
+  package then failed to lazy-load because `rstan`'s own precompiled
+  Windows binary DLL failed to load ("LoadLibrary failure: The
+  specified module could not be found"), most likely an ABI mismatch
+  between the RSPM binary and the runner's Rtools version. Fixed by
+  reinstalling `rstan` from source on the Windows CI job specifically.
+  Also added `src/Makevars.win` (lowering optimisation from `-O2` to
+  `-O1` and suppressing compiler warnings) and enlarged the Windows
+  runner's pagefile; these weren't the actual fix, but remain in place
+  as they reduce build time/noise and are otherwise harmless.
 
 ## Minor improvements
 
