### 2026-07-27 a2f9ff5
drop StanHeaders:::LdFlags() TBB linking, incompatible with RcppParallel 6.0.0 on Windows
With rstan now loading correctly (previous commit), Windows CI got
further and hit a second, distinct failure: our own compiled
EpiStrainDynamics.dll failed to load with the identical "LoadLibrary
failure" signature, this time for tbb.dll.

StanHeaders:::LdFlags() unconditionally returns
"-L<TBB_LIB> -Wl,-rpath,<TBB_LIB> -ltbb -ltbbmalloc", assuming
RcppParallel ships a dynamic tbb.dll/tbbmalloc.dll. RcppParallel 6.0.0
(installed on this runner) links TBB statically on Windows instead,
and -Wl,-rpath is a
[…]
