+++ articles/parameter-recovery.md
@@ -832,7 +832,7 @@ sessionInfo()
 #> [37] globals_0.19.1      scales_1.4.0        MASS_7.3-65        
 #> [40] anytime_0.3.13      cli_3.6.6           rmarkdown_2.31     
 #> [43] ragg_1.5.2          generics_0.1.4      otel_0.2.0         
-#> [46] RcppParallel_6.0.0  future.apply_1.20.2 cachem_1.1.0       
+#> [46] RcppParallel_6.1.0  future.apply_1.20.2 cachem_1.1.0       
 #> [49] rstan_2.32.7        splines_4.6.1       bayesplot_1.15.0   
 #> [52] parallel_4.6.1      matrixStats_1.5.0   vctrs_0.7.3        
 #> [55] hardhat_1.4.3       Matrix_1.7-5        jsonlite_2.0.0     
@@ -844,7 +844,7 @@ sessionInfo()
 #> [73] pillar_1.11.1       furrr_0.4.0         htmltools_0.5.9    
 #> [76] ipred_0.9-15        lava_1.9.2          R6_2.6.1           
 #> [79] textshaping_1.0.5   evaluate_1.0.5      lattice_0.22-9     
-#> [82] tsibble_1.2.0       bslib_0.11.0        rstantools_2.6.0   
+#> [82] tsibble_1.2.0       bslib_0.11.0        rstantools_2.7.0   
 #> [85] class_7.3-23        Rcpp_1.1.2          gridExtra_2.3.1    
 #> [88] prodlim_2026.03.11  xfun_0.60           fs_2.1.0           
 #> [91] zoo_1.8-15          pkgconfig_2.0.3
