### 2026-07-29 1b652d5
fix plot.incidence() hardcoding "influenza" in its y-axis label
The y-axis label was unconditionally set to "Modelled influenza cases"
regardless of the pathogen(s) actually being modelled, a leftover from
early development when the package's examples were influenza-specific.
Caught while using the sarscov2 dataset in the package's JSS submission
paper. Now reads the disease-agnostic "Modelled cases", consistent
with the other three plot() methods.

Closes #68.
