### 2026-08-03 55f7656
Restructure README and vignette into 4-phase framing to match paper structure
Reorganises the "Step 1/2/3/4" headers into Phase 1 (Preparing pathogen
data), Phase 2 (Configuring the model), Phase 3 (Fitting), and Phase 4
(Deriving epidemiological quantities), pulling pathogen_structure out
as its own phase and merging the old "Fit model"/"Check fit" sections.
No functional or code changes.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
