+++ vignettes/Using-EpiStrainDynamics.Rmd
@@ -13,25 +13,9 @@ vignette: >
 
 EpiStrainDynamics extends an existing statistical modelling framework capable of inferring the trends of up to two pathogens. The modeling framework has been extended here to handle any number of pathogens; fit to time series data of counts (eg, daily number of cases); incorporate influenza testing data in which the subtype for influenza A samples may be undetermined; account for day-of-the-week effects in daily data; include options for fitting penalized splines or random walks; support additional (optional) correlation structures in the parameters describing the smoothness of the penalized splines (or random walks); and account for additional (optional) sources of noise in the observation process.
 
-# Step 1: Construct model
+A full analysis proceeds through four phases, each producing an object consumed by the next: preparing the pathogen data, configuring the model, fitting it, and deriving epidemiological quantities.
 
-These modelling specifications are specified using the `construct_model` function. The correct stan model is then applied based on the specifications provided. `construct_model()` requires specification of the `method` and `pathogen_structure`, and allows the user to additionally specify four arguments related to model design: `smoothing_params`, `dispersion_params`, `pathogen_noise`, and `dow_effect`. We will break these down one by one.
-
-## Method
-
-EpiStrainDynamics has pre-compiled stan models that fit either with bayesian penalised splines or random walks. These are specified using the `method` argument of `construct_model()` as functions, either with `random_walk()` or `p_spline()`. The penalised spline model has two further options to specify: `spline_degree` is the polynomial degree of the individual spline segments used to construct the overall curve (must be a positive whole number) and `days_per_knot`, which is the number of days for each knot (must also be a positive whole number).
-
-So we may specify:
-```
-method = random_walk(),
-
-# OR
-
-method = p_spline(spline_degree = 3,  # example value for spline degree
-                  days_per_knot = 2)  # example value for days per knot
-```
-
-## Pathogen structure
+# Phase 1: Preparing pathogen data
 
 There are three main types of pathogen structure available to model: `single()`, `multiple()`, and `subtyped()`.
 These functions require the name of the dataset itself and the column names for different data elements. These can also ingest timeseries class objects, which would then make the `time` column specification optional. Any timeseries class input that is handled by the [`tsbox` package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html) is acceptable. The models are not currently equipped to handle missing values or irregular date sequences.
@@ -76,6 +60,26 @@ pathogen_structure = subtyped(
     'inf_B', 'other'
   )
 )
+```
+
+Each of these functions returns a pathogen structure object, which is then passed into `construct_model()` in Phase 2.
+
+# Phase 2: Configuring the model
+
[…]
