+++ README.Rmd
@@ -64,17 +64,30 @@ Each operating system has a different set up procedure. Windows users need to in
 
 ## Using `EpiStrainDynamics`
 
-Detailed instructions can be found on the [vignette](https://acefa-hubs.github.io/EpiStrainDynamics/articles/Using-EpiStrainDyamics.html). 
+Detailed instructions can be found on the [vignette](https://acefa-hubs.github.io/EpiStrainDynamics/articles/Using-EpiStrainDyamics.html).
 Here we provide a short overview.
 
-To build a model, you specify the model and pathogen structure with `construct_model()`, then fit it to your data with `fit_model()`. The steps below walk through this using the `sarscov2` data bundled with the package.
+A full analysis proceeds through four phases, each producing an object consumed by the next: preparing the pathogen data, configuring the model, fitting it, and deriving epidemiological quantities. The sections below walk through each phase using the `sarscov2` data bundled with the package.
 
-### Step 1: Construct model
+### Phase 1: Preparing pathogen data
 
-Modelling specifications are provided using the `construct_model()` function. The correct stan model is then applied based on the specifications provided. 
-`construct_model()` takes these arguments: `method`, `pathogen_structure`, `smoothing_params`, `dispersion_params`, `pathogen_noise`, and `dow_effect`. 
+There are three main types of pathogen structure available to model: `single()`, `multiple()`, and `subtyped()`.
+These functions require the name of the dataset itself and the column names for different data elements.
 
-#### `method`
+The `single()` pathogen structure is the simplest and models a single pathogen timeseries. The user must specify the dataset name, the name of the column with total case data, and the name of the column of time data.
+The `multiple()` pathogen structure allows modelling of different component pathogens. In addition to specifying the dataset, total case data, and time, the additional pathogens are specified as a vector of column names.
+The `subtyped()` pathogen structure enables additional complexity, specifically for an influenza modelling scenario, by allowing the user to incorporate testing data for influenza A subtypes.
+The user specifies columns containing the unsubtyped influenza A case count as well as the subtyped influenza A cases, and any additional pathogens to be modelled.
+See the vignette for further detail.
+
+Each of these functions returns a pathogen structure object, which is then passed into `construct_model()` in Phase 2.
+
+### Phase 2: Configuring the model
+
+Modelling specifications are provided using the `construct_model()` function. The correct stan model is then applied based on the specifications provided.
+`construct_model()` takes the pathogen structure object from Phase 1, plus these arguments describing the model: `method`, `smoothing_params`, `dispersion_params`, `pathogen_noise`, and `dow_effect`.
+
+#### Method
 
 EpiStrainDynamics has pre-compiled stan models that fit either with bayesian penalised splines or random walks. These are specified using the `method` argument of `construct_model()` as functions, either with `random_walk()` or `p_spline()`. The penalised spline model has two further options to specify: `spline_degree` is the polynomial degree of the individual spline segments used to construct the overall curve (must be a positive whole number) and `days_per_knot`, which is the number of days for each knot (must also be a positive whole number).
 
@@ -85,20 +98,9 @@ method = random_walk(),
 method = p_spline(spline_degree = 3, days_per_knot = 2)
 ```
 
-## Pathogen structure
+#### Smoothing parameters
 
-There are three main types of pathogen structure available to model: `single()`, `multiple()`, and `subtyped()`.
-These functions require the name of the dataset itself and the column names for different data elements.
-
-The `single()` pathogen structure is the simplest and models a single pathogen timeseries. The user must specify the dataset name, the name of the column with total case data, and the name of the column of time data. 
[…]
