+++ README.md
@@ -73,37 +73,13 @@ Detailed instructions can be found on the
 [vignette](https://acefa-hubs.github.io/EpiStrainDynamics/articles/Using-EpiStrainDyamics.html).
 Here we provide a short overview.
 
-To build a model, you specify the model and pathogen structure with
-`construct_model()`, then fit it to your data with `fit_model()`. The
-steps below walk through this using the `sarscov2` data bundled with the
+A full analysis proceeds through four phases, each producing an object
+consumed by the next: preparing the pathogen data, configuring the
+model, fitting it, and deriving epidemiological quantities. The sections
+below walk through each phase using the `sarscov2` data bundled with the
 package.
 
-### Step 1: Construct model
-
-Modelling specifications are provided using the `construct_model()`
-function. The correct stan model is then applied based on the
-specifications provided. `construct_model()` takes these arguments:
-`method`, `pathogen_structure`, `smoothing_params`, `dispersion_params`,
-`pathogen_noise`, and `dow_effect`.
-
-#### `method`
-
-EpiStrainDynamics has pre-compiled stan models that fit either with
-bayesian penalised splines or random walks. These are specified using
-the `method` argument of `construct_model()` as functions, either with
-`random_walk()` or `p_spline()`. The penalised spline model has two
-further options to specify: `spline_degree` is the polynomial degree of
-the individual spline segments used to construct the overall curve (must
-be a positive whole number) and `days_per_knot`, which is the number of
-days for each knot (must also be a positive whole number).
-
-So we may specify:
-
-    method = random_walk(),
-    # OR
-    method = p_spline(spline_degree = 3, days_per_knot = 2)
-
-## Pathogen structure
+### Phase 1: Preparing pathogen data
 
 There are three main types of pathogen structure available to model:
 `single()`, `multiple()`, and `subtyped()`. These functions require the
@@ -123,7 +99,36 @@ The user specifies columns containing the unsubtyped influenza A case
 count as well as the subtyped influenza A cases, and any additional
 pathogens to be modelled. See the vignette for further detail.
 
-## Smoothing parameters
+Each of these functions returns a pathogen structure object, which is
+then passed into `construct_model()` in Phase 2.
+
+### Phase 2: Configuring the model
+
+Modelling specifications are provided using the `construct_model()`
+function. The correct stan model is then applied based on the
+specifications provided. `construct_model()` takes the pathogen
+structure object from Phase 1, plus these arguments describing the
+model: `method`, `smoothing_params`, `dispersion_params`,
+`pathogen_noise`, and `dow_effect`.
+
+#### Method
+
+EpiStrainDynamics has pre-compiled stan models that fit either with
+bayesian penalised splines or random walks. These are specified using
+the `method` argument of `construct_model()` as functions, either with
+`random_walk()` or `p_spline()`. The penalised spline model has two
+further options to specify: `spline_degree` is the polynomial degree of
+the individual spline segments used to construct the overall curve (must
+be a positive whole number) and `days_per_knot`, which is the number of
+days for each knot (must also be a positive whole number).
+
+So we may specify:
+
+    method = random_walk(),
+    # OR
+    method = p_spline(spline_degree = 3, days_per_knot = 2)
+
+#### Smoothing parameters
 
 The argument `smoothing_params` allows users to modify the correlation
 structures in the parameters describing smoothness and to set related
@@ -148,7 +153,7 @@ pathogens are provided. As below:
       tau_sd = c(1, 1, 1, 1)
     )
 
-## Dispersion parameters
+#### Dispersion parameters
 
 The argument `dispersion_params` optionally allows users to set a prior
[…]
