### 2026-08-03 14bda22
Reorder construct_model() arguments: pathogen_structure before method
Swaps the first two arguments so pathogen_structure precedes method,
matching the logical order of building a model: prepare the pathogen
structure first, then choose how to model it. Updates the internal
validation order, all roxygen examples, README/vignette worked
examples, and test call sites (including several previously-positional
calls in test-construct-model.R that would otherwise have silently
swapped meaning).

This is a breaking change for any code calling construct_model()
positionally; calls using named argument
[…]
