+++ tests/testthat/helper-test-data.R
@@ -63,54 +63,54 @@ create_test_models <- function() {
   list(
     # Basic models with default parameters
     rw_single = construct_model(
-      method = random_walk(),
-      pathogen_structure = single_struct
+      pathogen_structure = single_struct,
+      method = random_walk()
     ),
     ps_single = construct_model(
-      method = p_spline(),
-      pathogen_structure = single_struct
+      pathogen_structure = single_struct,
+      method = p_spline()
     ),
     rw_multiple = construct_model(
-      method = random_walk(),
-      pathogen_structure = multiple_struct
+      pathogen_structure = multiple_struct,
+      method = random_walk()
     ),
     ps_multiple = construct_model(
-      method = p_spline(),
-      pathogen_structure = multiple_struct
+      pathogen_structure = multiple_struct,
+      method = p_spline()
     ),
     rw_subtyped = construct_model(
-      method = random_walk(),
-      pathogen_structure = subtyped_struct
+      pathogen_structure = subtyped_struct,
+      method = random_walk()
     ),
     ps_subtyped = construct_model(
-      method = p_spline(),
-      pathogen_structure = subtyped_struct
+      pathogen_structure = subtyped_struct,
+      method = p_spline()
     ),
 
     # Models with custom smoothing structure (shared)
     rw_multiple_shared_smooth = construct_model(
-      method = random_walk(),
       pathogen_structure = multiple_struct,
+      method = random_walk(),
       smoothing_params = smoothing_structure("shared", tau_mean = 0, tau_sd = 1)
     ),
     ps_multiple_shared_smooth = construct_model(
-      method = p_spline(),
       pathogen_structure = multiple_struct,
+      method = p_spline(),
       smoothing_params = smoothing_structure("shared", tau_mean = 0, tau_sd = 1)
     ),
 
     # Models with independent smoothing
     rw_multiple_indep_smooth = construct_model(
-      method = random_walk(),
       pathogen_structure = multiple_struct,
+      method = random_walk(),
       smoothing_params = smoothing_structure("independent",
         tau_mean = c(0, 0.1, 0.3, 0),
         tau_sd = rep(1, 4)
       )
     ),
     ps_subtyped_indep_smooth = construct_model(
-      method = p_spline(),
       pathogen_structure = subtyped_struct,
+      method = p_spline(),
       smoothing_params = smoothing_structure("independent",
         tau_mean = c(0, 0, 0.2, 0.1),
         tau_sd = rep(0.5, 4)
@@ -119,56 +119,56 @@ create_test_models <- function() {
 
     # Models with correlated smoothing
     rw_multiple_corr_smooth = construct_model(
-      method = random_walk(),
       pathogen_structure = multiple_struct,
+      method = random_walk(),
       smoothing_params = smoothing_structure("correlated")
     ),
     ps_subtyped_corr_smooth = construct_model(
-      method = p_spline(),
       pathogen_structure = subtyped_struct,
+      method = p_spline(),
       smoothing_params = smoothing_structure("correlated")
     ),
 
     # Models with custom dispersion parameters
     rw_multiple_custom_disp = construct_model(
-      method = random_walk(),
       pathogen_structure = multiple_struct,
+      method = random_walk(),
       dispersion_params = dispersion_structure(phi_mean = 2.0, phi_sd = 0.5)
     ),
     ps_subtyped_custom_disp = construct_model(
-      method = p_spline(),
       pathogen_structure = subtyped_struct,
+      method = p_spline(),
       dispersion_params = dispersion_structure(phi_mean = 1.5, phi_sd = 0.75)
     ),
 
     # Models with pathogen noise enabled
     rw_multiple_noise = construct_model(
[…]
