+++ R/plot.R
@@ -25,11 +25,11 @@
 #'
 #' @examplesIf interactive()
 #' mod <- construct_model(
-#'   method = random_walk(),
 #'   pathogen_structure = single(
 #'     case_timeseries = sarscov2$cases,
 #'     time = sarscov2$date
-#'   )
+#'   ),
+#'   method = random_walk()
 #' )
 #'
 #' fit <- fit_model(mod)
