### 2026-08-07 #763 @ropensci-review-bot
Approved! Thanks @smwindecker for submitting and @sangeetabhatia03, @joshwlambert for your reviews! :grin:
