### 2026-08-07 6f86656
cap viridis palette at end = 0.9 so lines don't fade to pale yellow
Also use black instead of an arbitrary viridis color for
plot.proportion() when only one numerator/denominator combo is
plotted, matching how the other plot() methods already handle a
single line.

Closes #71

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
