### 2026-08-15 86c6629
proportion() returns resolved numerator/denominator combinations; plot.proportion() shows a denominator subtitle
proportion() computed num_idx/denom_idx to select posterior samples for
the numerator and denominator, but discarded them once the calculation
was done. Nothing in the returned object recorded which pathogens made
up the denominator, so a plot built with a non-default
denominator_combination gave no way to tell what was excluded.

Adds $numerator_combination and $denominator_combination to the object
returned by proportion(), storing the resolved pathogen names. When
numerator_combi
[…]
