+++ NAMESPACE
@@ -52,6 +52,7 @@ importFrom(ggplot2,geom_point)
 importFrom(ggplot2,geom_ribbon)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,ggplot_build)
+importFrom(ggplot2,labs)
 importFrom(ggplot2,scale_colour_manual)
 importFrom(ggplot2,scale_y_continuous)
 importFrom(ggplot2,sec_axis)
