### 2026-08-15 274a57b
Remove influenzaA_ prefix from subtyped()'s timeseries arguments
subtyped()'s influenzaA_unsubtyped_timeseries and
influenzaA_subtyped_timeseries arguments implied the pathogen structure
only applied to influenza A subtypes, when it actually applies to any
pathogen with a combined, unsubtyped timeseries alongside partial
subtype data -- influenza A is just the most common example, not the
only one.

Renames the two arguments to unsubtyped_timeseries and
subtyped_timeseries, and the corresponding internal list key/variable
(pathogen_structure$data$influenzaA_subtyped, read by
construct_model())
[…]
