### 2026-08-15 d9b9a4c
Replace pkgdown site with redirect to docs.ropensci.org
Now that the package is under the ropensci org, docs.ropensci.org
builds and hosts the site automatically. Replace the old gh-pages
content with a redirect page, per
https://devguide.ropensci.org/redirect.html.
