+++ reference/subtyped.md
@@ -1,785 +1,0 @@
-# Subtyped pathogen structure
-
-Subtyped pathogen structure
-
-## Usage
-
-``` r
-subtyped(
-  data,
-  case_timeseries,
-  influenzaA_unsubtyped_timeseries,
-  influenzaA_subtyped_timeseries,
-  other_pathogen_timeseries,
-  time = NULL
-)
-```
-
-## Arguments
-
-- data:
-
-  dataframe containing columns with all relevant data, or a time series
-  object (ts, xts, zoo, tsibble, etc.)
-
-- case_timeseries:
-
-  Column name containing case counts. Must be numeric or a `units`
-  object from the units package.
-
-- influenzaA_unsubtyped_timeseries:
-
-  vector of column names with additional unsubtyped influenzaA case
-  count timeseries. Must be numeric or a `units` object from the units
-  package.
-
-- influenzaA_subtyped_timeseries:
-
-  vector of column names with additional subtyped influenzaA case count
-  timeseries. Must be numeric or a `units` object from the units
-  package.
-
-- other_pathogen_timeseries:
-
-  vector of column names with additional pathogen case count timeseries
-  to model. Must be numeric or a `units` object from the units package.
-
-- time:
-
-  name of column with time data. Required for non-time-series input
-  data. Flexible format - can be date, index, or others, accepted as
-  `index` identifiers in the `tsibble` time format. Optional when `data`
-  is a time series class object (ts, mts, xts, zoo, zooreg, tsibble) as
-  the time index will be automatically detected.
-
-## Value
-
-named list including pathogen_structure, pathogen_names, and data of
-class `EpiStrainDynamics.pathogen_structure`
-
-## See also
-
-Other pathogen_structure:
-[`multiple()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/multiple.md),
-[`single()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/single.md)
-
-## Examples
-
-``` r
-# Using a data frame
-subtyped(
-  data = influenza,
-  case_timeseries = "ili",
-  influenzaA_unsubtyped_timeseries = "inf_A",
-  influenzaA_subtyped_timeseries = c("inf_H3N2", "inf_H1N1"),
-  other_pathogen_timeseries = c("inf_B", "other"),
-  time = "week"
-)
-#> $pathogen_structure
-#> [1] "subtyped"
-#> 
-#> $pathogen_names
-#> [1] "inf_H3N2" "inf_H1N1" "inf_B"    "other"   
-#> 
-#> $validated_tsbl
-#> # A tsibble: 426 x 7 [7D]
-#>    week       case_timeseries inf_A inf_H3N2 inf_H1N1 inf_B other
-#>    <date>               <int> <int>    <int>    <int> <int> <int>
-#>  1 2012-01-02              29     4        4        0     3   278
-#>  2 2012-01-09              38     4        4        0     1   219
-#>  3 2012-01-16              42    13       13        0     8   238
-#>  4 2012-01-23              58     6        3        2     5   187
-#>  5 2012-01-30              45     3        3        0     2   228
-#>  6 2012-02-06              60     1        1        0     7   274
-#>  7 2012-02-13              30     2        0        2    10   278
-#>  8 2012-02-20              52     4        3        1     9   266
-#>  9 2012-02-27              33     6        4        2    13   327
-#> 10 2012-03-05              26     4        4        0    10   324
-#> # ℹ 416 more rows
-#> 
-#> $data
-#> $data$case_timeseries
-#>   [1]  29  38  42  58  45  60  30  52  33  26  75  65  71  92  71  75  73  39
-#>  [19]  57  94 140 144 218 234 176 265 370 389 469 465 434 496 441 399 359 255
[…]
