+++ articles/index.md
@@ -1,10 +1,0 @@
-# Articles
-
-### Statistical Properties
-
-- [Using
-  EpiStrainDynamics](https://acefa-hubs.github.io/EpiStrainDynamics/articles/Using-EpiStrainDynamics.md):
-- [Algorithmic Scaling and Statistical
-  Properties](https://acefa-hubs.github.io/EpiStrainDynamics/articles/algorithmic-scaling.md):
-- [Parameter Recovery
-  Tests](https://acefa-hubs.github.io/EpiStrainDynamics/articles/parameter-recovery.md):
