+++ reference/multiple.md
@@ -1,1116 +1,0 @@
-# Multiple pathogen structure
-
-Multiple pathogen structure
-
-## Usage
-
-``` r
-multiple(data, case_timeseries, component_pathogen_timeseries, time = NULL)
-```
-
-## Arguments
-
-- data:
-
-  dataframe containing columns with all relevant data, or a time series
-  object (ts, xts, zoo, tsibble, etc.)
-
-- case_timeseries:
-
-  Column name containing case counts. Must be numeric or a `units`
-  object from the units package.
-
-- component_pathogen_timeseries:
-
-  vector of column names with additional pathogen case count timeseries
-  to model. Must be numeric or a `units` object from the units package.
-
-- time:
-
-  Column name with time data. Required for non-time-series input data.
-  Flexible format - can be date, index, or others, accepted as `index`
-  identifiers in the `tsibble` time format. Optional when `data` is a
-  time series class object (ts, mts, xts, zoo, zooreg, tsibble) as the
-  time index will be automatically detected.
-
-## Value
-
-named list including pathogen_structure, pathogen_names, and data of
-class `EpiStrainDynamics.pathogen_structure`
-
-## See also
-
-Other pathogen_structure:
-[`single()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/single.md),
-[`subtyped()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/subtyped.md)
-
-## Examples
-
-``` r
-# Using a data frame
-multiple(
-  data = sarscov2,
-  case_timeseries = "cases",
-  component_pathogen_timeseries = c("alpha", "delta", "omicron", "other"),
-  time = "date"
-)
-#> $pathogen_structure
-#> [1] "multiple"
-#> 
-#> $pathogen_names
-#> [1] "alpha"   "delta"   "omicron" "other"  
-#> 
-#> $validated_tsbl
-#> # A tsibble: 830 x 6 [1D]
-#>    date       case_timeseries alpha delta omicron other
-#>    <date>               <dbl> <int> <int>   <int> <int>
-#>  1 2020-09-23            7096     0     0       0    10
-#>  2 2020-09-24            7682     0     0       0     7
-#>  3 2020-09-25            7318     0     0       0     9
-#>  4 2020-09-26            6810     0     0       0    15
-#>  5 2020-09-27            7023     0     0       0     5
-#>  6 2020-09-28            9983     0     0       0     6
-#>  7 2020-09-29           10366     0     0       0    23
-#>  8 2020-09-30           12717     0     0       0    54
-#>  9 2020-10-01           13357     0     0       0    32
-#> 10 2020-10-02           13854     0     0       0    15
-#> # ℹ 820 more rows
-#> 
-#> $data
-#> $data$case_timeseries
-#>   [1]   7096   7682   7318   6810   7023   9983  10366  12717  13357  13854
-#>  [11]  11667  11953  16753  17250  18573  18425  15893  12648  12202  19671
-#>  [21]  19116  19920  18606  17915  14882  14311  25872  25613  25714  23475
-#>  [31]  21694  16434  15895  26807  24351  23901  23614  22890  16673  15970
-#>  [41]  31925  25943  24027  24063  23659  18766  20244  31510  27789  27757
-#>  [51]  24826  24388  18223  16233  26952  23488  21076  18041  17100  12533
-#>  [61]  11895  19024  16679  16814  14761  14554  11086  10852  18198  16684
-#>  [71]  16676  15860  15834  12893  13731  21584  21065  21909  22916  24390
-#>  [81]  19588  22266  35197  34378  35427  34437  36419  25589  33066  48192
[…]
