### 2026-08-15 5afba87
Post-rOpenSci-transfer housekeeping: version bump, links, CoC, pkgdown
The repo transferred from acefa-hubs to the ropensci org after the
rOpenSci statistical software peer review (gold badge, review #763).

- Bump version 0.0.1.9000 -> 0.1.0 and retitle the NEWS.md heading to
  reflect that this is a release, not a dev snapshot
- Add the two rOpenSci reviewers (Sangeeta Bhatia, Joshua Lambert) to
  Authors@R with role "rev"
- Point all repo/docs links (DESCRIPTION, README, codemeta.json,
  CITATION, CONTRIBUTING.md, and the piggyback CI fixture-download
  refs) at ropensci/EpiStrainDynamics a
[…]
