+++ NEWS.md
@@ -1,4 +1,4 @@
-# EpiStrainDynamics (development version)
+# EpiStrainDynamics 0.1.0
 
 ## Breaking changes
 
