### 2026-08-15 11ce158
Add rOpenSci Software Peer Review status badge to README
